refactor: rename view-edit to view-editor
