+++ src/lib/layouts/rest/views/index.ts
@@ -3,6 +3,6 @@ export * from './view-instructions';
 export * from './view-import-from-file';
 export * from './view-collections';
 export * from './view-select-collections';
-export * from './view-edit';
+export * from './view-editor';
 export * from './view-response';
 export * from './view-rest';
