--- src/lib/layouts/rest/views/view-edit/index.ts
@@ -1,3 +1,0 @@
-import { default as ViewEdit } from './view-edit.layout.svelte';
-export { ViewEdit };
-export default ViewEdit;
