+++ src/lib/layouts/rest/views/view-editor/index.ts
@@ -1,0 +1,3 @@
+import { default as ViewEditor } from './view-editor.layout.svelte';
+export { ViewEditor };
+export default ViewEditor;
